[UPDATE] recent render & auto copy
