+++ js/copy/copyBtn.mjs
@@ -1,7 +1,7 @@
 import Doc from "../service/doc.mjs";
 
 function copyBtn() {
-  const btn = Doc.find(".copy_btn");
+  const btn = Doc.find("#copy_btn");
   btn.style.opacity = "0.5";
   setTimeout(function() {
     btn.style.opacity = "1.0";
