+++ js/settings/sizeSetting.js
@@ -3,6 +3,7 @@ import { sizeSetting } from "./localVariable.mjs";
 import setGrid from "../setGrid.mjs";
 
 const sizeBtns = Doc.findAll(".sizeBtn");
+const groups = Doc.findAll(".emoji-span-container");
 
 function setSizeBtn() {
   setCss();
@@ -30,7 +31,7 @@ function addSizeEvent() {
     sizeBtns[0].classList.add("active");
     sizeBtns[1].classList.remove("active");
     sizeBtns[2].classList.remove("active");
-    setGrid(15);
+    setGrid(15, groups);
   });
 
   sizeBtns[1].addEventListener("click", () => {
@@ -39,7 +40,7 @@ function addSizeEvent() {
     sizeBtns[0].classList.remove("active");
     sizeBtns[1].classList.add("active");
     sizeBtns[2].classList.remove("active");
-    setGrid(10);
+    setGrid(10, groups);
   });
 
   sizeBtns[2].addEventListener("click", () => {
@@ -48,7 +49,7 @@ function addSizeEvent() {
     sizeBtns[0].classList.remove("active");
     sizeBtns[1].classList.remove("active");
     sizeBtns[2].classList.add("active");
-    setGrid(5);
+    setGrid(5, groups);
   });
 }
